requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
langchain>=1.0.0
langchain-classic>=1.0.0      # create_retrieval_chain, create_stuff_documents_chain
langchain-core>=1.0.0         # ChatPromptTemplate, MessagesPlaceholder, HumanMessage
langchain-text-splitters>=0.2.0
langchain-anthropic>=0.1.15
langchain-community>=0.2.0    # PyPDFLoader, FAISS
langchain-huggingface>=0.0.3  # HuggingFaceEmbeddings
faiss-cpu==1.7.4              # wheel precompilado; 1.8+ requiere swig
sentence-transformers<3.0.0   # 3.x requiere transformers 5.x → torch>=2.4
numpy<2                       # faiss-cpu 1.7.4 fue compilado con numpy 1.x
pypdf>=4.2.0
